Enable operating on values using Python statements.
With this change, all uknown commands will be treated as Python
expressions and/or statements. In essence, this change embeds a
limited Python interpreter into graftr.

If the current working directory in graftr is `foo`, all value nodes
in `foo` will be exposed as local variables for the interpreter.
Those value nodes can be reassigned or modified in-place with standard
Python statements. The `torch` and `np` modules are also available to
the user.

For example:

```
$ graftr /path/to/checkpoint.pt
Checkpoint loaded from /path/to/checkpoint.pt.
Type help or ? to list commands.

> print('Hello world')
Hello world
> cd model/linear1
> ls
weight
bias
> cat weight
tensor([[ 0.0021, -0.0485,  0.0316,  ..., -0.0321,  0.0583,  0.0442],
        [ 0.0386, -0.4908, -0.5002,  ..., -0.2382,  0.0724,  0.0051],
        [-0.1682, -0.4633, -0.3734,  ...,  0.0914,  0.0299,  0.0086],
        ...,
        [-0.0181,  0.0035, -0.0767,  ...,  0.0661, -0.1513,  0.1622],
        [ 0.0188, -0.0192, -0.0006,  ..., -0.0079,  0.0099, -0.0567],
        [-0.0310,  0.1529,  0.4278,  ...,  0.0347, -0.0990,  0.0009]],
       device='cuda:0')
> weight[0]
tensor([ 0.0021, -0.0485,  0.0316,  ..., -0.0321,  0.0583,  0.0442],
       device='cuda:0')
> weight[0] = torch.zeros_like(weight[0])
> weight[0]
tensor([0., 0., 0.,  ..., 0., 0., 0.], device='cuda:0')
> cat weight
tensor([[ 0.0000,  0.0000,  0.0000,  ...,  0.0000,  0.0000,  0.0000],
        [ 0.0386, -0.4908, -0.5002,  ..., -0.2382,  0.0724,  0.0051],
        [-0.1682, -0.4633, -0.3734,  ...,  0.0914,  0.0299,  0.0086],
        ...,
        [-0.0181,  0.0035, -0.0767,  ...,  0.0661, -0.1513,  0.1622],
        [ 0.0188, -0.0192, -0.0006,  ..., -0.0079,  0.0099, -0.0567],
        [-0.0310,  0.1529,  0.4278,  ...,  0.0347, -0.0990,  0.0009]],
       device='cuda:0')
```